+++ packages/eslint-config-wpcalypso/index.js
@@ -1,5 +1,3 @@
-/** @format */
-
 module.exports = {
 	env: {
 		es6: true,
@@ -110,7 +108,14 @@ module.exports = {
 		// Assumed by default with Babel
 		strict: [ 2, 'never' ],
 		'template-curly-spacing': [ 2, 'always' ],
-		'valid-jsdoc': [ 2, { requireReturn: false } ],
+		'valid-jsdoc': [
+			2,
+			{
+				requireParamType: false,
+				requireReturn: false,
+				requireReturnType: false,
+			},
+		],
 		'wpcalypso/i18n-ellipsis': 2,
 		'wpcalypso/i18n-no-collapsible-whitespace': 2,
 		'wpcalypso/i18n-no-variables': 2,